optimized usage of interface: FormData
